fix(demo): feedback 2 gst
